backend/requirements.txt
fastapi>=0.98.0
uvicorn[standard]
motor
python-jose
passlib[bcrypt]
python-dotenv
pydantic>=2.0,<3.0
pydantic-settings
pydantic[email]
PyJWT
requests
hugging_hub
